Consolidate openshift-apiserver and -controller-manager configs
